+++ CHANGELOG.md
@@ -7,6 +7,9 @@ This project **DOES NOT** adhere to [Semantic Versioning](http://semver.org/spec
 ## [Unreleased]
 [Unreleased]: https://bitbucket.org/atlassian/db-replica/branches/compare/master%0Drelease-0.1.8
 
+### Fixed
+- Assign deletes to the main connection for `executeQuery` calls
+
 ### Removed
 - `api.circuitbreaker.DualConnectionException`